--- src/seedsigner/extras/dev.py
@@ -1,4 +1,0 @@
-import os
-from typing import Final
-
-SIMULATE_MODE: Final = os.environ.get("LOCAL_DEV") == "true"
