--- src/seedsigner/extras/local_display.py
@@ -1,11 +1,0 @@
-class ST7789:
-    def __init__(self, width = 240, height = 240):
-        self.width = width
-        self.height = height
-
-    def show_image(self, image, x, y):
-        with open('display.bmp', 'wb') as f:
-            f.write(image.tobytes())
-
-    def invert(self, enabled: bool = True):
-        pass
