+++ src/seedsigner/hardware/displays/display_driver.py
@@ -1,6 +1,3 @@
-from seedsigner.extras.dev import SIMULATE_MODE
-
-
 DISPLAY_TYPE__ST7789 = "st7789"
 DISPLAY_TYPE__ILI9341 = "ili9341"
 DISPLAY_TYPE__ILI9486 = "ili9486"
@@ -22,17 +19,11 @@ class DisplayDriver:
                 # TODO: For now the original ST7789 driver has to be used for 240x240.
                 # The mpy version below renders incorrectly (almost like each row of pixels
                 # is one pixel short, so the entire screen exhibits a diagonal skew).
-                if not SIMULATE_MODE:
-                    from seedsigner.hardware.displays.ST7789 import ST7789
-                else:
-                    from seedsigner.extras.local_display import ST7789
+                from seedsigner.hardware.displays.ST7789 import ST7789
                 self.display = ST7789()
 
             elif width == 320:
-                if not SIMULATE_MODE:
-                    from seedsigner.hardware.displays.st7789_mpy import ST7789
-                else:
-                    from seedsigner.extras.local_display import ST7789
+                from seedsigner.hardware.displays.st7789_mpy import ST7789
                 # Have to swap width and height; screen is natively 240x320
                 self.display = ST7789(width=height, height=width)
